requirement.txt
streamlit
pandas
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
chromadb
sentence-transformers